+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.0.5
+
+* [Fixed] Invalid `micropackage/acf-block-creator/config` hookname
+* [Added] Filters for block tamplate and style file paths
+
 ## 1.0.4
 
 * [Fixed] block setup fields behavior